// src/api/generated/schemas/apiExceptionResponse.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * WorkLog API
 * API para gerenciamento de tickets e interações de suporte ao cliente. Autenticação via cookie HttpOnly `worklog_access` emitido pelo endpoint `POST /worklog/auth/login` (o browser anexa o cookie automaticamente em chamadas subsequentes).
 * OpenAPI spec version: v1.0.0
 */
import type { ApiExceptionResponseFieldErrors } from './apiExceptionResponseFieldErrors';
import type { MediaType } from './mediaType';

export interface ApiExceptionResponse {
  mediaType?: MediaType;
  timestamp?: string;
  status?: number;
  error?: string;
  message?: string;
  path?: string;
  fieldErrors?: ApiExceptionResponseFieldErrors;
}